handle and log pull request status update errors
Previously, when the GitHubService attempted to update the pull request
status, any unsuccessful API response would cause an exception to be
raised. As a result, if the GitHub web token was authorized to _read_
repository contents but not to _write/push_ to the repository (as is
required to update commit statuses), the processing for a webhook PR
event would prematurely fail with an uncaught exception and,
consequently, no scan would be started, nor would a result summary
comment be posted to the PR, even though both actions could otherwise
be performed with the read-only permissions.

This commit resolves this issue by catching and logging status API
response errors, whether caused by missing repository permissions/
scopes, or by any other issue causing the API request to fail.