Fix for issue where GitLabIssueTracker can sometimes find the wrong project when searching for details. (#245)
The GitLab search API uses the project name being searched for as a token. This means that any project that contains that token in its name will be returned. The existing code assumed that only one project would be returned, because an exact match search would occure.

So, searching for: dvna-0429-1 would return dvna-0429-1 but also projects with names dvna-0429-11 and dvna-0429-10.

The fix examins all returned projects and ensure that only the correct one is picked.

Co-authored-by: Jeff Armstrong <jeff.armstrong@checkmarx.com>