Allow Jira issue summary to be specified by configuration
Previously, the format of the Jira issue summary (for SAST issues) was
hard-coded. With this change, it can be configured via the
jira.sast-issue-summary-format and
jira.sast-issue-summary-branch-format configuration
properties. The formatting mechanism is similar to that used to
construct filenames for the CSV, JSON and XML backends: certain
variable fields, enclosed in square brackets are replaced with the
appropriat values. The following fields are supported:

- Branch
- Filename
- Postfix
- Prefix
- Project
- Severity
- Vulnerability