Replace text even when the replacement string is null or empty
The original logic was the same as that used in the
SanitizingFilenameFormatter class (from which the implementation was
copied). However, this does not make sense for the Jira summary as some
placeholders may legitimately be null or empty (e,g., the issue prefix
and postfix properties).

Now, if the actual value is null or empty, the placeholder is replaced
with an empty string.