Fix the issue for CLI scan mode
Create the issue in Jira with Label as SAST scan and SCA scan.
Filter the issue on bases of type of scan (SCA or SAST) and Update/Open/Close the issue for which scan is initiated for the project.
If scan is initiated for both SCA and SAST then issues for both type of scan will be update for that project.