Support suppression of code snippets for Jira tickets (#976)
* Support suppression of code snippets for Jira tickets

A new property has been added to the JiraProperties class:
suppressCodeSnippets. This property is a list of vulnerability names
for which code snippets should not be generated when creating a Jira
ticket.

The rationale behind this is that some code snippets can contain
sensitive information (e.g., hard-coded passwords).

* Cosmetic tweak: suppress entire code snippet section

The prior commit suppressed only the code snippet itself. This commit
also suppresses the dividing line and the line number label for the
snippet.